TA and sampled analysis almost ready to go for the full data set.  The first analysis breaks the data set into subsets along sample size and TA duration lines, runs a random forest classifier on each with 10 fold CV, and populates a data frame with accuracy, kappa, and ROC curve values for analysis.
I need to go back to the census and pure sampled data and get all the comparable values
into a common data frame so I can compare the ROC curves and Kappa values across ALL
of the data sets, but I think I'm seeing a drastic decline in Kappa here.  That's
from a small subsample for testing, however.